support Float32 structs

+++ src/io.jl
@@ -781,7 +781,7 @@ function prepare_writer(
             time_units,
             maxlayers,
             sizeinmetres;
-            float_type = Float64,
+            float_type = Float,
         )
     else
         ds_outstate = nothing
@@ -1124,7 +1124,7 @@ end
 
 "Read a rating curve from CSV into a NamedTuple of vectors"
 function read_sh_csv(path)
-    data, header = readdlm(path, ',', Float64, header = true)
+    data, header = readdlm(path, ',', Float, header = true)
     names = vec(uppercase.(header))
     idx_h = findfirst(==("H"), names)
     idx_s = findfirst(==("S"), names)
@@ -1138,11 +1138,11 @@ end
 
 "Read a specific storage curve from CSV into a NamedTuple of vectors"
 function read_hq_csv(path)
-    data = readdlm(path, ',', Float64, skipstart = 1)
+    data = readdlm(path, ',', Float, skipstart = 1)
     # Q is a matrix with 365 columns, one for each day in the year
     return (H = data[:, 1], Q = data[:, 2:end])
 end
 
 # these represent the type of the rating curve and specific storage data
-const SH = NamedTuple{(:H, :S),Tuple{Vector{Float64},Vector{Float64}}}
-const HQ = NamedTuple{(:H, :Q),Tuple{Vector{Float64},Matrix{Float64}}}
+const SH = NamedTuple{(:H, :S),Tuple{Vector{Float},Vector{Float}}}
+const HQ = NamedTuple{(:H, :Q),Tuple{Vector{Float},Matrix{Float}}}
